# Avoid rendering the whole event to get only the provider name