--- a/.chloggen/avoid_rendering_whole_event_to_obtain_provider_name_windowseventlogreceiver.yaml
+++ b/.chloggen/avoid_rendering_whole_event_to_obtain_provider_name_windowseventlogreceiver.yaml
@@ -10,7 +10,7 @@ component: windowseventlogreceiver
 note: Avoid rendering the whole event to obtain the provider name
 
 # Mandatory: One or more tracking issues related to the change. You can use the PR number here if no issue exists.
-issues: []
+issues: [34755]
 
 # (Optional) One or more lines of additional information to render under the primary note.
 # These lines will be padded with 2 spaces and then inserted directly into the document.